Replace namespace FuseSource with Stomp in order to get a clean cut and starting point for new client version.
