+++ examples/transactions.php
@@ -19,7 +19,7 @@ require __DIR__.'/../vendor/autoload.php';
  */
 
 // include a library
-use FuseSource\Stomp\Stomp;
+use Stomp\Stomp;
 // make a connection
 $con = new Stomp("tcp://localhost:61613");
 // connect
