+++ src/Stomp/Exception/UnexpectedResponseException.php
@@ -1,7 +1,8 @@
 <?php
-namespace FuseSource\Stomp\Exception;
+namespace Stomp\Exception;
 
-use Exception;
+use Stomp\Frame;
+
 /**
  *
  * Copyright 2005-2006 The Apache Software Foundation
@@ -22,21 +23,38 @@ use Exception;
 /* vim: set expandtab tabstop=3 shiftwidth=3: */
 
 /**
- * Base exception for all special stomp exceptions.
+ * Exception that occurs, when a frame / response was received that was not expected at this moment.
+ *
  *
  * @package Stomp
+ * @author Jens Radtke <swefl.oss@fin-sn.de>
  */
-class StompException extends Exception
+class UnexpectedResponseException extends StompException
 {
+    /**
+     *
+     * @var Frame
+     */
+    private $_frame;
 
     /**
-     * Stomp server error details
      *
-     * @deprecated use getMessage()
-     * @return string
+     * @param Frame $frame
+     * @param string $expectedInfo
      */
-    public function getDetails()
+    function __construct(Frame $frame, $expectedInfo)
     {
-        return $this->getMessage();
+        $this->_frame = $frame;
+        parent::__construct(sprintf('Unexpected response received. %s', $expectedInfo));
     }
+
+    /**
+     *
+     * @return Frame
+     */
+    public function getFrame()
+    {
+        return $this->_frame;
+    }
+
 }
